Tweak proactive scheduling
Use only a third of the items (1) and only look 1 minute ahead (2).

(1). We still scale too agressively and there is a big stair climbing
down.
(2). Scaling is fast, so we can scale just one minute in advance and
take better advantage of the 5 minute cool down period before we start
scaling down.