Add buffer instances for API
Right now the API instances would start from a baselive of 4 and would
not scale up until the requests per minute threshold was crossed
(i.e. 1500 * number of instances). This approach does not work for
spiked traffic as we need to wait a full minute to get the new request
count from ELB.

This PR gives the option to configure a different number of extra
instances per environment for the notify-api app, to act as a buffer
until we scale them up.