Increased notify-delivery-worker-receipts max to 30
After tests sending 1 million email an hour, the number of receipt
workers did not keep up with the demand meaning that the functional
tests failed due to the delay in processing the receipt. Increased the
max number of receipt workers to 30 to ensure there are enough workers
to cope with the demand. Bother 20 and 25 instances were tested and
showed a significant increase in queue size over a sustained run i.e.
15-30 mins.

- Updated the notify-delivery-worker-receipts to use the v_high value
- Added a V_HIGH value of 30
- Updated manifest to include export the value