Added Template Preview to the Autoscaler
This is an incremental change to cope with the extra demand on template
preview for PDF letters. Further improvements will be made to the app
itself. Currently Template preview is not scaled and a single instance
is deployed. This is not enough to cope with the demand of a batch of
letters sent via the API or via CSV and causes all but the
first batch (currently 10) to fail. These will retry and due to the more
staggered nature of the calls, template preview copes a little better.

This change is to ensure that if there is a large batch (100s) that the
retries do not get out of hand, so scale the template preview app in a
similar way to the other apps based on the queue size.

Only scaled to 10 instances as template preview has a larger footprint
that other apps at 2GB of memory and their is a finite amount of memory
that we have (also the cost of running hose extra instances), so created
a 'medium' instance count of 10. That still only gives us a max
throughput of approx 50 pdf letters a second. However as letters only get
sent once a day this should be more than enough to keep up with any surge
in calls.