Don't bind unused hosted-graphite PaaS user service
hosted-graphite service contains a JSON data blob with `statsd_prefix`,
but it's not actually used by the application. Instead, STASD_PREFIX
is read from the PaaS application environment variables.

We can unbind the service so that it can be removed now that it's not
used by any other apps on PaaS.