Add support for simple scheduled scaling
When the Autoscaler starts, it will load `schedule.yml` that
contains per-app configuration for scheduling its scaling.

The format of the file is:

```
name-of-the-app:
  <workdays|weekends>:
      - HH:MM-HH:MM
      - HH:MM-HH:MM
      - ...
name-of-another-app:
  <workdays|weekends>:
      - HH:MM-HH:MM
      - ...
```

The check of whether or not `now()` is within the time ranges defined in
`schedule.yml` happens every `SCHEDULE_INTERVAL` like every other check
and will scale the app to at least `SCHEDULED_SCALE_FACTOR * max_instance_count`
**unless** the other metrics of the app requires it to scale to a higher
number of instances, in which case it will scale to that.

The schedule file contained in this PR only defines scheduling rules for
the `notify-delivery-worker-sender`