map route (even though we won't use it for anything)
v3-zdt-push maps a route for web processes, even though we've specified
`no-route: true` in the manifest.

web processes need a route, and if a routes dict isn't specified
they'll create `{{app_name}}.cloudapps.digital`, which causes problems
when we try and push the same app to two separate environments.

so even though we've got no web process listening for incoming
connections, we need to define a route just to stop conflicts across
envs.

you can define worker processes instead (which this probably should be),
but cloudfoundry will create a web process and map the route to it
anyway. You can scale down the web process to 0 and unmap the route,
but I can't figure out a good way to get cf to do this for you and I'd
rather not add more sequential steps to the deployment process at the
moment. it also appears that the memory/disk quota etc in the manifest
only affect the web process.