Make scaling timezone-aware
Yesterday morning there was a spike on email traffic between 08:30 and
08:40 that were not delivered within 10 seconds. A quick investigation
showed that the autoscaler did not increase the number of instances
at 08:00 according to its schedule but, since we switched to BST, it
scales at 09:00 instead.

This commit addresses this inconsistency by using `pytz` to get
timezone-aware time objects.

`pytz` is a dependency of `notifications-utils` so in order to avoid
version clashes I'm just explicitly listing it as a dependency without
version.