Set max number of API instances to 25
We add a new variable to do this as we don't need to scale the
workers that high too.

This was based on an incident today of a group of 502s and we
believe increasing the number of APIs may help address similar
traffic spikes in the future however as we never got down to the
root cause we are mostly hoping this change will be beneficial.