Add autoscaling for template-preview-celery
This will give us between 2 and 10 instances in production, scaling on
the template-preview queue (`sanitise-letter-tasks`) and on the two
queues that precompiled letters pass through before being put on the
template preview queue.

The delivery apps typically have a threshold for the SqsScaler of 250
and 11 workers. Template preview celery has 5 workers and its task takes
a little longer to run, so the threshold for scaling has been made
lower.