scale to max if surge queue ever has content
 The Surge Queue is a queue in AWS ELB - if none of our apps accept an
incoming connection, it'll queue up to 1024 connections that our nginx
(really our api) instances aren't able to accept. We shouldn't ever
have anything in this queue - so if it's nonzero we should scale up.

In the short term this metric is stored per second, but we should get
the max val from the last minute anyway to be safe.

Scale up to max just to be safe.