scale worker instances based on throughput as well as queue length
currently we scale based on how many items are in queues waiting to be
processed. However, this doesn't distinguish between "lots of items
being processed quickly, nothing building up in the queue" (where we
shouldn't downscale) and "no items at all coming in to the queue" (where
it's okay to downscale).

To solve this, add a check of the NumberOfMessagesReceived metric, which
sums up the number of messages entering each queue. In the same way as
we scale the web apps based off the ELB, look at the last five minutes
and get the max value per minute, then scale based off that.

If there is a large amount of data coming in, AND we have a large amount
in the queue already, we want to scale for both of those things, so add
together the existing desired instance count and the new one based on
messages received.