Allow receipts to scale up to 25 in production
We haven't been able to handle the load in the SES queue, it's gone up
to 80000 twice today. The impact of this is that our provider smoke
tests fail and that reply-to emails can't be set for users.

I decided to do this individually for the receipts app, rather than
bumping the `max_instance_count_high` as this would have raised the
number of instances for 6 other apps that don't need it and we shouldn't
add too many apps if we aren't sure we need them to reduce the risk of
running out of DB connections (we are currently running at 3.5k out of
5k).