Bump min instances to 18 for sender-worker
Regularly through the day, we get large spikes in email traffic. At this
point, we would usually be running around 8 instances. When we get a
large spike of items in the email sending queue, we will need to scale
up to more like 18 instances. It takes us a minute or two to do this
before we then drain down that queue. This means lots of emails have
waited for around a minute to get processed which affects our overall
performance stats for sending under 10 seconds.

Data to back this up is regular spikes in the email sender queue with
lots of items waiting for around a minute to get processed:
https://grafana-paas.cloudapps.digital/d/9RTLOhyZk/notify-queues?orgId=2&from=now-30d&to=now

By running minimum 18 instances of the sending worker during daytime
hours we should remove the issue of waiting for scaling when a spike of
emails comes in. We can then see how this affects our performance over
the next week and if it solves the problems we've seen with roughly only
87% of emails being delivered within 10 seconds.

18 instances was chosen for two reasons
- by not running at max 20, we can still see the autoscaler working in
  practice
- based on data from grafana showing most spikes of instance numbers go
  up to about 15-18 instances so this should remove most of those needs
to scale. Data to back this up:
https://grafana-paas.cloudapps.digital/d/_GlGBNbmk/notify-apps?orgId=2&var-space=production&var-app=notify-delivery-worker-sender&from=1586270939631&to=1586570840458&fullscreen&panelId=22