Run save-api-notifications worker at min 18 instances in prod
We did similar for the sender worker in
https://github.com/alphagov/notifications-paas-autoscaler/pull/89

We want to also do this for the save-api-notifications worker because we
are seeing spikes in the queue it pulls from and then it has to scale up
and deal with that. The time to scale means we aren't sending out those
notifications within 10 seconds. This should help us get closer to
meeting that target.

For data see the following and choose to view only the
`livesave-api-email-tasks` queue:
https://grafana-paas.cloudapps.digital/d/9RTLOhyZk/notify-queues?orgId=2&from=1586778806524&to=1586950242673