update thresholds for template-preview-celery
set tasks_per_worker_per_minute to 25 (as seen when under load), rather
than the default of 1000.

rename threshold to allowed_queue_backlog_per_worker so it makes more
sense in context in the config file, and also set it to 25 so that we
add a worker per minutes worth of work on the backlog.