reduce notify-template-preview scaling
set tasks per worker per minute to 300. This is based on 5 gunicorn
worker threads processing 1 task per second each over 60 seconds. Over
the last week, we've had 62% of requests to /print.pdf (the only
endpoint hit from the create-letters-pdf-tasks worker) under 1 second,
31% between 1 and 2 seconds, and 8% over 2 seconds.

Note: this doesn't take into account the baseline of admin traffic,
which is approximately 36% of the traffic, and has a similar response
time spread. But we shouldn't scale the app for admin traffic based on
the queue workers. In an ideal world we would scale based on traffic to
template preview itself as well as the SQS queues.

set allowed queue backlog per worker to 100. If the queue has grown to
100 items, create another instance to help deal with the backlog