Scale reporting app to 3
This change will mean that we run 3 instances during midnight to 4am
(when our reporting tasks run) and then outside of those hours it will
scale back down to it's minimum of a single instance.

Note, midnight to 4am was chosen as the reporting tasks currently kick
off at 00:15. We expect them to have finished by 3am but I've kept it
scaled until 4am just in case something had massively overrun (although
we hope not) and to therefore avoid killing an instance that still might
be executing a task.

Note, we don't anticipate any problems with BST. Our celery beat
schedule works on BST and so does the autoscaler so they are matched (no
situation where it's UTC midnight and BTC 11pm and tasks are kicked
off).