Remove unneeded sqs scaling
The number is definitely very wrong for us. We would usually only put 18
tasks on this a day so we will never hit 250 and scale up.

We potentially could pick a more appropriate number for the scaling in
the future but I'm not trying to solve that problem now as instead we've
done it based on a schedule which assumes we know when our tasks will
arrive (just after midnight).