Allow template preview celery to scale up to 35 instances
Last week had the smoke tests fail because a letter took more than
a minute to make it through template preview. This was caused by
template preview celery being underscaled. This was concluded by looking
at the number of items being put on the queue vs the number of items
being pulled off it. The number of items being put on was higher than
the ability to process them all for about 10 minutes, causing a delay
and the smoke tests to fail.

35 instances was chosen as we at one point we were receiving almost
double the number items onto the queue as we were pulling off it so we
are almost doubling it from 20 to 35.

It is debatable whether we should keep scaling this app horizontally
(more instances) as this costs money. At some point we should work out
what our time threshold is for a letter to work it's through antivirus
and template preview and scale to keep letter processing under that
threshold. At the moment it's implied to be a minute based on smoke
tests but I don't know if that is a correct assumption or just happens
to be that way.