Set delivery-jobs-worker to 25 instances min and max
We want to stop it scaling down for the moment because we have lots of
big CSVs to process and want it to be ready for them when they come.

We also raise it from 20 to 25 max as it's not keeping up with the
numbers of CSVs that are being uploaded.

There could be knock on effects further down the change so we aren't
bumping up the max number of instances too high in one go and will need
to monitor.