Allow receipts worker to scale to 30 instances
At high loads, we are seeing the queue building up and smoke tests fail
because we aren't able to process tasks of this queue quick enough. We
allow it to scale to 30 instead of 35 max. This should help but not be
too big a jump that we risk a big increase in DB connections.