Change minimum number of job workers back to 2
it was previously changed to run at minimum of 25 (same as the max) so
that we weren't scaling and we were ready for large amounts of load
during a busy weekend.

This now sets it back to what the original minimum was pre weekend of 2
instances (the same value as `MIN_INSTANCE_COUNT_LOW`).