Update tests due to bump in cloudfoundry-client
In the previous commit, we bumped cloudfoundry-client from 1.10.0 to
1.14.1. This introduced a change in what argument type an
`InvalidStatusCode` should take. It should now take a `HTTPStatus`
object.

This change can be seen in
https://github.com/cloudfoundry-community/cf-python-client/commit/08e8f28328504164aa4ce6bf90a5254aa03bcacc
and
https://github.com/cloudfoundry-community/cf-python-client/blob/a9635e9d9ae4aa10d1037e5040d9dbf7aad4cef3/main/cloudfoundry_client/errors.py#L10.

This required changing one of our tests to match the argument type an
`InvalidStatusCode` and change our logging assertion as we now will get
a slightly different log line due to the change of the exception.