Scale research worker to max 30 instances
It currently is allowed to scale up to 10 instances

When a team load tested using test API keys at 200/s we saw that 10
instances couldn't handle that and that a large queue built up.

We know have a request to load test at 650/s. Therefore if we triple the
number of instances our performance should get no worse.

There is an argument that we should let this scale above 30 workers but
I don't want to give too many instances to start with because they use
DB connections and whilst it's likely to be OK, I would rather still see
some delays for test notification processing then risk changing our
system too much.