Scale research worker higher to support 600rps
We've had a short load test at 600rps using test keys. This led to
delays and we build up a queue.

They are now going to test for more like an hour at this rate and the
queue will continue to grow. If that is not acceptable then we should
scale this worker up.

Based on a graph of how many tasks we were taking from the queue when
running 10 instances, it was about 10K per minute. At a rate of 600rps
we put about 50K tasks on a queue per minute. Therefore, 50 instances
should roughly handle 600rps.