Add metrics for tasks pulled from queue
We previously had these on our grafana dashboard and they were useful to
look at during incidents/debugging. They were replaced by metrics for
how many tasks are being put onto the queue in
https://github.com/alphagov/notifications-paas-autoscaler/pull/101

Now, however we want these back on our grafana dashboard. This gives us
a few options of how to do this
- set up grafana with an AWS Cloudwatch integration
- create a new app to handle grabbing metrics from AWS Cloudwatch to be
  scraped by prometheus
- wack the metrics in this app because it has very similar code for the
  number of metrics in the queue and being put onto the queue

Decided to go for the third option due to speed and ease. In the long
run we may want to go down one of the other routes.

Will require a separate PR to add the statsd mapping for this new
metric.

I've mostly kept the code as similar as possible to calculating the
throughput of tasks put onto the queues. This does give us a tiny bit of
redundant code (we don't really care about the past 5 minutes of data
for example), but I think the duplicated code makes it easier to
understand/maintain/debug than having more snowflaked code in here just
for this quick fix of providing metrics.