Set instance count for broadcasts worker
Note, it is always set as 2 in production because it's highly unlikely
we will get many tasks put on the queue at a time (potentially 4 or 5 at
a time).

We could have done this by just setting the number of instances in the
API manifest to keep it at 2 instances always, but we added it here as
it gives us metrics for queue length/tasks picked up etc in grafana
automatically. Therefore, the threshold choice of 50 is fairly abitrary
as I don't expect it to ever scale based on this (if it could as it's
already at it's max).