Move bootstrap tasks into the Makefile
This is more consistent with how we run all other tasks. Note that
the virtual env setup is not generally applicable, and developers
of this repo should follow the guidance in the README.

For some reason the bootstrap script wasn't installing the test
dependencies. This is done separately by Concourse [1], which didn't
use the bootstrap.sh script.

[1]: https://github.com/alphagov/notifications-aws/blob/1c836ac68aa5dcb85a87cdb1bad8e79846cdafd3/concourse/templates/macros.yml.j2#L197