Bump max research workers from 50 to 60
We are currently seeing the research worker developing queues that
they are not keeping up with.

We are currently running on 50 research workers and they appear to
be handling just under 50K tasks per minute. We are putting just
over 50K tasks on the queue per minute so hopefully this 20% increase
in workers will hit the sweet spot to catch up and bring down the queue.

Note, the research worker does touch the DB. Currently the DB health
is looking good, in particular DB connections, such that I don't think
this is a particularly risky change.