Add `freeze-requirements` Makefile rule
When trying to install the requirements, pip's dependency resolver was
backtracking and attempting to install every version of awscli
https://pip.pypa.io/en/stable/topics/dependency-resolution/#backtracking.
This meant that the app couldn't be deployed because Concourse ran out
of space.

We can fix this by taking the approach we use in other apps of freezing
requirements.

I've had to use older versions of the `cloudfoundry-client` and `pyyaml`
in order to avoid conflicts. The `cloudfoundry-client` specifies an exact
version of `pyyaml` to install - later versions tried to install `pyyaml 6.0`
which conflicts with `awscli`.