Add scaling for `notify-delivery-worker-sender-letters`
This adds the `SqsScaler` to the new
`notify-delivery-worker-sender-letters` app. Since this isn't running
yet we are not sure how long each task will take but can change the
number of instances following the planned load tests. The main thing we
want to avoid here is exceeding DVLA's rate limit of 50 requests per
second. Using a maximum of 5 instances in production should avoid going
over the limit, but we may find that we can have more instances after
testing.