Scale up periodic worker by 4 instances
https://github.com/alphagov/notifications-api/pull/3918

That PR is going to put 4 long running tasks on the periodic workers.
So we don't take up all the periodic worker doing those tasks
(which would block them picking up other things like sending
scheduled jobs), we need to increase the minimum and max number
of instances on the assumption that 4 of them will be running the
above task all the time.

It does make the assumption that the task above will take less than
24 hours (otherwise, we will go from having 4 running tasks
to having 8 running tasks).