Scale down broadcasts worker to 1 in all environments
One, this will save us money as it is no longer used.

Two, it will reduce the throughput on a task I've put on the
broadcast worker which may be having a small impact on production
api response times. This will slow it down and reduce that until
we are happy to scale it back up more.